ADDED: Some casts to fix some Visual Studio 2008 warnings
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@876 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfPagesTree.cpp
@@ -373,7 +373,7 @@ void PdfPagesTree::InsertPages( int inAfterIndex,
 #ifdef _MSC_VER		
         // cannot access empty deque, 
         // TODO:perhaps better move up in next function ??!!
-        if ( i >= m_deqPageObjs.size())
+        if ( i >= static_cast<int>(m_deqPageObjs.size()) )
             m_deqPageObjs.push_back( NULL );
 #endif
